refactor: move code related functions to a separate util file
