+++ frontend/src/components/PropsEditor.vue
@@ -78,7 +78,8 @@ import EmptyState from "@/components/EmptyState.vue"
 import Block from "@/utils/block"
 
 import InlineInput from "@/components/InlineInput.vue"
-import { isDynamicValue, isObjectEmpty } from "@/utils/helpers"
+import { isObjectEmpty } from "@/utils/helpers"
+import { isDynamicValue } from "@/utils/code"
 import IconButton from "@/components/IconButton.vue"
 import Link2 from "~icons/lucide/link-2"
 import Link2Off from "~icons/lucide/link-2-off"
