+++ frontend/src/stores/appStore.ts
@@ -3,7 +3,8 @@ import { ref, watch, computed, type WatchStopHandle } from "vue"
 import { studioPageResources } from "@/data/studioResources"
 import { studioVariables } from "@/data/studioVariables"
 import { studioWatchers } from "@/data/studioWatchers"
-import { getInitialVariableValue, getNewResource, executeUserScript } from "@/utils/helpers"
+import { getInitialVariableValue } from "@/utils/helpers"
+import { getNewResource, executeUserScript } from "@/utils/code"
 import app_router from "@/router/app_router"
 
 import type { Resource } from "@/types/Studio/StudioResource"
